Extract statistics panel into a separate component
