+++ src/pages/ProfilePage.js
@@ -1,12 +1,9 @@
-import React, { useState, useContext } from "react";
+import React, { useState } from "react";
 import { Redirect } from "react-router-dom";
 
 import Container from "@material-ui/core/Container";
 import Divider from "@material-ui/core/Divider";
 import Grid from "@material-ui/core/Grid";
-import List from "@material-ui/core/List";
-import ListItem from "@material-ui/core/ListItem";
-import ListItemText from "@material-ui/core/ListItemText";
 import { makeStyles } from "@material-ui/core/styles";
 import Paper from "@material-ui/core/Paper";
 import Table from "@material-ui/core/Table";
@@ -19,17 +16,10 @@ import Typography from "@material-ui/core/Typography";
 import EqualizerIcon from "@material-ui/icons/Equalizer";
 import StarIcon from "@material-ui/icons/Star";
 
-import { Pie } from "react-chartjs-2";
-
-import { UserContext } from "../context";
 import ProfileName from "../components/ProfileName";
+import UserStatistics from "./UserStatistics";
 
 const useStyles = makeStyles((theme) => ({
-  statisticsPanel: {
-    background: theme.palette.background.default,
-    padding: theme.spacing(2),
-    borderRadius: 4,
-  },
   mainGrid: {
     "--table-height": "400px", // responsive variable
     [theme.breakpoints.up("sm")]: {
@@ -67,29 +57,10 @@ const useStyles = makeStyles((theme) => ({
       display: "none",
     },
   },
-  usernamePanel: { display: "flex", flexDirection: "column" },
 }));
 
 function ProfilePage({ match }) {
   const [redirect, setRedirect] = useState(null);
-
-  // Fake pie chart data
-  const data = {
-    datasets: [
-      {
-        data: [35, 100],
-        backgroundColor: ["#81c784", "#e57373"],
-      },
-    ],
-    labels: ["Games won", "Games lost"],
-  };
-  const options = {
-    maintainAspectRatio: false,
-    responsive: true,
-    legend: {
-      position: "bottom",
-    },
-  };
 
   const classes = useStyles();
 
@@ -113,44 +84,7 @@ function ProfilePage({ match }) {
               <Typography variant="overline"> Statistics</Typography>
               <EqualizerIcon />
             </div>
-            <Grid container direction="row" className={classes.statisticsPanel}>
-              <Grid item xs={12} sm={6} md={6} className={classes.pie}>
-                <Pie data={data} options={options} />
-              </Grid>
-              <Grid item xs={12} sm={6} md={6}>
-                <List dense>
-                  <ListItem>
-                    <Typography variant="body2">Finished games: 135</Typography>
-                  </ListItem>
-                  <ListItem>
-                    <ListItemText>
-                      <Typography variant="body2">Total sets: 2025</Typography>
-                    </ListItemText>
-                  </ListItem>
-                  <ListItem>
-                    <ListItemText>
-                      <Typography variant="body2">
-                        Average sets per game: 15
-                      </Typography>
-                    </ListItemText>
-                  </ListItem>
-                  <ListItem>
-                    <ListItemText>
-                      <Typography variant="body2">
-                        Fastest game won: 5 min 30 sec
-                      </Typography>
-                    </ListItemText>
-                  </ListItem>
-                  <ListItem>
-                    <ListItemText>
-                      <Typography variant="body2">
-                        Average game length: 4 min
-                      </Typography>
-                    </ListItemText>
-                  </ListItem>
-                </List>
-              </Grid>
-            </Grid>
+            <UserStatistics></UserStatistics>
           </Grid>
 
           <Grid item xs={12} sm={12} md={12}>
